LPS-75465 Move com.liferay.blogs.recent.bloggers.web.constants to com.liferay.blogs.recent.bloggers.constants
