+++ modules/apps/collaboration/blogs/blogs-layout-prototype/build.gradle
@@ -10,4 +10,5 @@ dependencies {
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:collaboration:blogs:blogs-api")
+	provided project(":apps:collaboration:blogs:blogs-recent-bloggers-api")
 }
