+++ modules/apps/collaboration/blogs/blogs-layout-prototype/src/main/java/com/liferay/blogs/layout/prototype/internal/instance/lifecycle/AddLayoutPrototypePortalInstanceLifecycleListener.java
@@ -17,7 +17,7 @@ package com.liferay.blogs.layout.prototype.internal.instance.lifecycle;
 import com.liferay.asset.tags.navigation.constants.AssetTagsNavigationPortletKeys;
 import com.liferay.blogs.constants.BlogsPortletKeys;
 import com.liferay.blogs.model.BlogsEntry;
-import com.liferay.blogs.recent.bloggers.web.constants.RecentBloggersPortletKeys;
+import com.liferay.blogs.recent.bloggers.constants.RecentBloggersPortletKeys;
 import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
 import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
