+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-api/src/main/resources/com/liferay/blogs/recent/bloggers/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
