+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-test/build.gradle
@@ -1,10 +1,10 @@
 dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.arquillian.extension.junit.bridge", version: "1.0.8"
-	testIntegrationCompile group: "com.liferay", name: "com.liferay.blogs.recent.bloggers.web", version: "1.0.0"
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.registry.api", version: "1.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.6.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
 	testIntegrationCompile project(":apps:collaboration:blogs:blogs-api")
+	testIntegrationCompile project(":apps:collaboration:blogs:blogs-recent-bloggers-api")
 	testIntegrationCompile project(":apps:collaboration:blogs:blogs-service")
 	testIntegrationCompile project(":apps:collaboration:blogs:blogs-test-util")
 	testIntegrationCompile project(":apps:web-experience:export-import:export-import-api")
