+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-test/src/testIntegration/java/com/liferay/blogs/recent/bloggers/internal/exportimport/portlet/preferences/processor/test/RecentBloggersExportImportPortletPreferencesProcessorTest.java
@@ -15,7 +15,7 @@
 package com.liferay.blogs.recent.bloggers.internal.exportimport.portlet.preferences.processor.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.blogs.recent.bloggers.web.constants.RecentBloggersPortletKeys;
+import com.liferay.blogs.recent.bloggers.constants.RecentBloggersPortletKeys;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.portlet.preferences.processor.ExportImportPortletPreferencesProcessor;
 import com.liferay.exportimport.test.util.ExportImportTestUtil;
