+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-web/build.gradle
@@ -14,5 +14,6 @@ dependencies {
 	provided group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:collaboration:blogs:blogs-api")
+	provided project(":apps:collaboration:blogs:blogs-recent-bloggers-api")
 	provided project(":apps:web-experience:export-import:export-import-api")
 }
