+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-web/src/main/java/com/liferay/blogs/recent/bloggers/web/internal/exportimport/data/handler/RecentBloggersPortletDataHandler.java
@@ -14,7 +14,7 @@
 
 package com.liferay.blogs.recent.bloggers.web.internal.exportimport.data.handler;
 
-import com.liferay.blogs.recent.bloggers.web.constants.RecentBloggersPortletKeys;
+import com.liferay.blogs.recent.bloggers.constants.RecentBloggersPortletKeys;
 import com.liferay.exportimport.kernel.lar.DataLevel;
 import com.liferay.exportimport.kernel.lar.DefaultConfigurationPortletDataHandler;
 import com.liferay.exportimport.kernel.lar.PortletDataHandler;
