+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-web/src/main/java/com/liferay/blogs/recent/bloggers/web/internal/portlet/action/ViewMVCRenderCommand.java
@@ -14,7 +14,7 @@
 
 package com.liferay.blogs.recent.bloggers.web.internal.portlet.action;
 
-import com.liferay.blogs.recent.bloggers.web.constants.RecentBloggersPortletKeys;
+import com.liferay.blogs.recent.bloggers.constants.RecentBloggersPortletKeys;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 
 import javax.portlet.RenderRequest;
